Move Ethereum JSON RPC API in silkrpc/commands Rename silkrpc::json::eth as silkrpc::json
